Parte 2/text_analytics/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2
django-appconf==1.0.5
django-compressor==4.3.1
djangorestframework==3.14.0
docopt==0.6.2
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.5.0/es_core_news_sm-3.5.0-py3-none-any.whl#sha256=13b909bac0de5f5338f4327b14a9cc9736cf51f07131a3cfcb688dbeb8608bdd
executing==1.2.0
flatbuffers==23.3.3
fonttools==4.39.3
fpdf2==2.7.4
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.12.0
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
multimethod==1.9.1
murmurhash==1.0.9
mysqlclient==2.1.1
networkx==3.1
nltk==3.8.1
num2words==0.5.12
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
pandas-profiling==3.6.6
parso==0.8.3
pathy==0.10.1
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==9.5.0
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==4.22.3
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
Pygments==2.15.1
PyMySQL==1.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
rcssmin==1.1.1
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rjsmin==1.2.1
rsa==4.9
scikeras==0.10.0
scikit-keras==0.2.0
scikit-learn==1.2.2
scipy==1.9.3
seaborn==0.12.2
six==1.16.0
smart-open==6.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.4
srsly==2.4.6
stack-data==0.6.2
statsmodels==0.13.5
tangled-up-in-unicode==0.2.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thinc==8.1.9
threadpoolctl==3.1.0
tqdm==4.64.1
traitlets==5.9.0
typeguard==2.13.3
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
visions==0.7.5
wasabi==1.1.1
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
ydata-profiling==4.1.2
zipp==3.15.0
